+++ src/hooks/forms/index.ts
@@ -1,3 +1,4 @@
 export * from "./useConfirmationForm";
+export * from "./useMutationForm";
 export * from "./usePostForm";
 export type * from "./types";
